Fix warnings

+++ compiler/protobuf/qml/util.cpp
@@ -37,7 +37,7 @@ std::string camelize(const std::string& name) {
 }
 
 std::string capitalizeAll(std::string name) {
-  for (int i = 0; i < name.size(); ++i) {
+  for (size_t i = 0; i < name.size(); ++i) {
     auto c = name[i];
     name[i] = is_small_char(c) ? capitalize(c) : c;
   }
